vps/requirements.txt
fastapi==0.104.1
fastapi-mail==1.4.1
fastapi-sso==0.9.1
gunicorn==21.2.0
uvicorn==0.24.0.post1
pydantic==2.5.2
SQLAlchemy==2.0.23
SQLAlchemy-Utils==0.41.1
asyncpg==0.29.0
passlib==1.7.4
alembic==1.13.0
psycopg2==2.9.9
python-multipart==0.0.6
python-jose==3.3.0
openai==1.3.7
aiohttp==3.9.1
sse-starlette==1.8.2
starlette==0.27.0
Authlib==1.2.1
httpx==0.25.2
itsdangerous==2.1.2
stripe==7.7.0
celery==5.3.6
soundfile==0.12.1
librosa==0.10.1
numpy==1.24.4
python-dateutil==2.8.2
requests==2.31.0
pydub==0.25.1
Jinja2==3.1.2
pytest==7.4.3
sqladmin==0.16.0
elevenlabs==0.2.26
redis==5.0.1
mailchimp-transactional==1.0.50
bcrypt==4.1.2
pyairtable==2.2.1
cent==5.0.0